\name{createBlockPredGrid}
\alias{createBlockPredGrid}
\title{Creates grid of prediction points within a boundary}
\usage{
  createBlockPredGrid(bnd)
}
\arguments{
  \item{bnd}{an outline of a single polygon of sp Class
  SpatialPolygon}
}
\value{
  a prediction grid as a SpatialPoints class.
}
\description{
  Creates grid of prediction points within boundary
}
\author{
  Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}

