\name{OzoneCA}
\alias{OzoneCA}
\docType{data}
\title{
Data Folder Containing Shape Files for California Ozone Concentrations
}
\description{
  This data set came with ESRI Geostatistical Analyst, and was used in their manuals.
}
\details{
	There are two shapefiles, ca_outline and ca_ozone in the /inst/rawdata/airPolluteCA folder

}
\source{
  ESRI
}


