\name{print.summary.spCountSamp}
\alias{print.summary.spCountSamp}
\title{Printing a Summary of a Spatial Linear Mixed Model Fits}
\usage{
  \method{print}{summary.spCountSamp} (x, ...)
}
\arguments{
  \item{object}{a summary of a fitted splmm object}
}
\value{
  the function \code{print.summary.splmm} the list of
  summary statistics of the fitted spatial linear mixed
  model
}
\description{
  method for class summary.splmm
}

